terraform/cloudrunjob/cloudrunimage/requirements.txt
numpy==1.23.5
pandas==1.5.3
pyyaml==6.0
sqlalchemy==1.4.46
google-cloud-bigquery==3.10.0
pyarrow==12.0.0
mysql-connector-python==8.1.0